refactor CreateImageIndex task; improve logging; log tip for enabling conrainerd
